Remove ResamplingSpecification from forecast
